added ssh function for public instane to upload files and run commands

+++ smichaud/utils/create_key_pair.py
@@ -1,13 +1,17 @@
 import boto3
 import os
+from pathlib import Path
 def generate_key_pair(ec2_client, key_pair_name, out_path = "temp"):
     # Generate a key pair
-    
+    key_pair_path = Path(os.path.join(out_path, f'{key_pair_name}.pem'))
+    if key_pair_path.exists():
+        print(f"Key pair '{key_pair_name}' already exists.")
+        return key_pair_path
     response = ec2_client.create_key_pair(KeyName=key_pair_name)
 
     # Save the private key to a file
     private_key = response['KeyMaterial']
-    key_pair_path = os.path.join(out_path, f'{key_pair_name}.pem')
+    Path(out_path).mkdir(exist_ok=True)
     with open(key_pair_path, 'w') as key_file:
         key_file.write(private_key)
